+++ lib/const/regex_pattern.dart
@@ -1,4 +1,6 @@
 const String kEmailPtrn =
     r'^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$';
 
-const String kHaveDigitPtrn = '.*[0-9].*';
+const String kContainDigitPtrn = r'.*[0-9].*';
+
+const String kOnlyDigitPtrn = r'^-?[0-9]*$';